+++ GHEtool/VariableClasses/PipeData/MultipleUTube.py
@@ -385,6 +385,10 @@ class MultipleUTube(_PipeData):
             # factor 4 instead of 2 since here R_b is used instead of R_1b
             R_a = 4 * R_b * R12 / (4 * R_b + R12)
 
+            # store r_b and R_a
+            self._R_b = R_b
+            self._R_a = R_a
+
             # convert to the effective borehole resistance using the constant borehole wall boundary condition
             r_v = borehole.H / (flow_rate_data.mfr_borehole(**kwargs, fluid_data=fluid_data) * fluid_data.cp(
                 **kwargs) / self.number_of_pipes)
@@ -512,6 +516,10 @@ class MultipleUTube(_PipeData):
             else:
                 raise NotImplementedError(
                     'Explicit models are only implemented for double U probes are only implemented for orders 0 an 1.')
+        # store r_b and R_a
+        self._R_b = R_b
+        self._R_a = R_a
+
         r_v = borehole.H / (flow_rate_data.mfr_borehole(**kwargs, fluid_data=fluid_data) * fluid_data.cp(
             **kwargs) / self.number_of_pipes)
         n = r_v / (self.number_of_pipes * R_b * R_a) ** 0.5